+++ demo/src/client/pages/Home/Home.tsx
@@ -2,7 +2,7 @@ import { Page } from '../../components';
 
 export function Home() {
   return (
-    <Page title="Home">
+    <Page title="Home" view="general">
       <p>Hi</p>
     </Page>
   );