+++ packages/core/src/api/meta/initialize.ts
@@ -1,6 +1,6 @@
 import type { Config } from '../../config';
 import type { InternalLogger } from '../../internalLogger';
-import type { Meta, Metas, MetaSession, MetaUser } from '../../metas';
+import type { Meta, Metas } from '../../metas';
 import type { Transports } from '../../transports';
 import type { UnpatchedConsole } from '../../unpatchedConsole';
 import type { MetaAPI } from './types';
@@ -16,8 +16,9 @@ export function initializeMetaAPI(
 
   let metaSession: Partial<Meta> | undefined = undefined;
   let metaUser: Partial<Meta> | undefined = undefined;
+  let metaView: Partial<Meta> | undefined = undefined;
 
-  const setUser = (user?: MetaUser) => {
+  const setUser: MetaAPI['setUser'] = (user) => {
     if (metaUser) {
       metas.remove(metaUser);
     }
@@ -29,7 +30,7 @@ export function initializeMetaAPI(
     metas.add(metaUser);
   };
 
-  const setSession = (session?: MetaSession) => {
+  const setSession: MetaAPI['setSession'] = (session) => {
     if (metaSession) {
       metas.remove(metaSession);
     }
@@ -43,11 +44,27 @@ export function initializeMetaAPI(
 
   const getSession: MetaAPI['getSession'] = () => metas.value.session;
 
+  const setView: MetaAPI['setView'] = (view) => {
+    if (metaView) {
+      metas.remove(metaView);
+    }
+
+    metaView = {
+      view,
+    };
+
+    metas.add(metaView);
+  };
+
+  const getView: MetaAPI['getView'] = () => metas.value.view;
+
   return {
     setUser,
-    resetUser: setUser,
+    resetUser: setUser as MetaAPI['resetUser'],
     setSession,
-    resetSession: setSession,
+    resetSession: setSession as MetaAPI['resetSession'],
     getSession,
+    setView,
+    getView,
   };
 }