+++ packages/core/src/api/meta/types.ts
@@ -1,9 +1,11 @@
-import type { MetaSession, MetaUser } from '../../metas';
+import type { MetaSession, MetaUser, MetaView } from '../../metas';
 
 export interface MetaAPI {
-  setUser: (user: MetaUser) => void;
+  setUser: (user?: MetaUser | undefined) => void;
   resetUser: () => void;
-  setSession: (session: MetaSession) => void;
+  setSession: (session?: MetaSession | undefined) => void;
   resetSession: () => void;
   getSession: () => MetaSession | undefined;
+  setView: (view?: MetaView | undefined) => void;
+  getView: () => MetaView | undefined;
 }