+++ packages/core/src/metas/index.ts
@@ -15,4 +15,5 @@ export type {
   MetaSDKIntegration,
   MetaSession,
   MetaUser,
+  MetaView,
 } from './types';