+++ packages/web-sdk/src/metas/index.ts
@@ -3,3 +3,7 @@ export { browserMeta } from './browser';
 export { defaultMetas } from './const';
 
 export { pageMeta } from './page';
+
+export { createSession } from './session';
+
+export { defaultViewMeta } from './view';