+++ packages/web-sdk/src/metas/view/index.ts
@@ -1,0 +1,1 @@
+export { defaultViewMeta } from './const';